NAS-116088: Refactoring GroupQuotaFormComponent and UserQuotaFormComponent
