+++ src/app/pages/storage/volumes/datasets/dataset-quotas/dataset-quotas-grouplist/group-quota-form/group-quota-form.component.scss
@@ -1,0 +1,3 @@
+.form-actions {
+  margin: 8px 10px;
+}
